Changed Individual to be a typeclass
I'm not sure if this will work out, but it was extremely fun working
out how to convince the compiler to do what I want.

This forces people who want to call "evolve" to implement their join
and breed functions using only the Individual typeclass functions. As a
result, I can hide from the user the fact that I'm caching the fitness
score in a record.