Rewrite of the generic GA code to actually work
Not sure how I feel about the functional deps on the Individual and
EvolutionStrategy type classes, but I needed it to make the methods not using
the monad (fitness, joinGens) to not be ambiguous.

Also in this commit, evolve changed from taking the number of generations to
returning an infinite list... seems like a more elegant API.  We'll see how it
plays out.  Sidenote: it turns out to be ~10-15% faster than the recursive
version. ¯\_(ツ)_/¯