refactor evolve to use Pipes rather than janky, accidentally working list stuff
This has been on my todo list for entirly too long.  The list stuff that was
here before worked, but kind of by accident.  Each slot in the list was not
actually dependent on the slot before it, so it only worked because I was
ignoring everything but the 25th slot of each list, and constructing a new list
when starting the next batch.

This also has the advantage of not recomputing the fitnesses for the first
generation of each batch since interally it still uses the cached
individuals.